created hook for password validation to avoid code duplication
