+++ app/src/pages/outside/common/usePasswordValidation/index.js
@@ -1,0 +1,1 @@
+export { usePasswordValidation } from './usePasswordValidation';
